// frontend/src/app/master/models/generalMetrics.ts
export interface generalMetrics {
  id: number | null;                   
  generalMetricsName: string;
  description: string | null;
  startDate:    string | null;
  endDate: string | null;
  remarks: string | null;
  active?: boolean;                   
  status: number;                     
  createdBy?: string |null;                 
  modifiedBy?: string|null;   
}             

  

 